Add files via upload
Me and my friends set out on a task to find the best time to find a relationship and the time to be careful of breakups.  
We collected our data from surveying around the UIC campus gathering relationship trends and analyzing them to unfold the discoveries behind love and heartbreak. We further implemented a ML model to find how likely your relationship is going to work out based on the relationship trends. 

Link to our google survey: 
https://docs.google.com/forms/d/e/1FAIpQLSc6SRVPG1Jc-ORkP96eZF0a9MH2VfSkxNZ1_8aAgWxZeUYFzQ/viewform